CI: Configure `fail-fast: true`, to save resources
Usually, when a single job fails, it does not make sense to continue
running the other jobs in the same context.

When debugging dedicated issues which only manifest on specific versions
of Python, or on specific operating system environments, it is easy to
toggle that option back to `fail-fast: false`.