NwpIndex: Persist original Dataset's schema as netCDF and JSON
Because the engine breaks out the actual data from the Dataset into an
ironArray, and its metadata will get lost, it is needed to collect the
schema information about coordinates and dimensions and store it
separately from the `.iarr` data.

This will allow to reconstruct the original Dataset as close as
possible when reading the data back.

Currently, two files, `schema.nc`, and `schema.json` will be stored
within the corresponding `.iarr` folder.